Reworked plugin dependency handling.
Because plugin manager must be noticed about adding/removing a plugin
to/from "used by" list, it is now done directly through plugin manager
which has the dependency, not through given plugin metadata.

The only exception are static plugins with no associated manager, for
them the "used by" list is managed manually (and it's okay, because
there is no plugin manager to receive the notice). The manager is now
also disconnected from static plugins on deletion, so another manager
can take over them in future.

In addition fixed crash in case when one plugin depends on another
non-static in the same manager. If on manager destruction the dependency
is destroyed before the plugin which depends on it, the depending plugin
would access these nonexistent data and cause crash. Now the plugin data
are destroyed in separate cycle to avoid that. It still needs unload
with proper dependency handling, though.

Also removed check for duplicates when adding plugin to "used by" list,
because on the contrary it was not checked when removing the plugin.
Also, if there would be two plugins with the same name, it is needed to
have both of them in "used by" list for dependency handling on unload to
work correctly.